Add raft_digest to generate IDs

+++ src/election.c
@@ -38,8 +38,8 @@ struct followerOrCandidateState *getFollowerOrCandidateState(struct raft *r)
 void electionResetTimer(struct raft *r)
 {
     struct followerOrCandidateState *state = getFollowerOrCandidateState(r);
-    unsigned timeout =
-        r->io->random(r->io, r->election_timeout, 2 * r->election_timeout);
+    unsigned timeout = (unsigned)r->io->random(r->io, (int)r->election_timeout,
+                                               2 * (int)r->election_timeout);
     assert(timeout >= r->election_timeout);
     assert(timeout <= r->election_timeout * 2);
     state->randomized_election_timeout = timeout;
