+++ src/entry.h
@@ -5,7 +5,7 @@
 
 /* Release all memory associated with the given entries, including the array
  * itself. The entries are supposed to belong to one or more batches. */
-void entryBatchesDestroy(struct raft_entry *entries, unsigned n);
+void entryBatchesDestroy(struct raft_entry *entries, size_t n);
 
 /* Create a copy of a log entry, including its data. */
 int entryCopy(const struct raft_entry *src, struct raft_entry *dst);
@@ -14,6 +14,6 @@ int entryCopy(const struct raft_entry *src, struct raft_entry *dst);
  * including their data. */
 int entryBatchCopy(const struct raft_entry *src,
                    struct raft_entry **dst,
-                   unsigned n);
+                   size_t n);
 
 #endif /* ENTRY_H */
